# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: StudyFlow Local API (Draft)
  version: 0.1.0
  description: |
    Draft OpenAPI spec for the local backend. Scope is limited to core routes
    implemented under `server/src/routes/`. This spec will evolve and expand.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.studyflow.local
    description: Local or staging gateway
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            example:
              value:
                type: https://httpstatuses.com/401
                title: Request Error
                status: 401
                detail: Missing or invalid bearer token
                instance: /api/courses
    TooManyRequests:
      description: Too Many Requests
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            example:
              value:
                type: https://httpstatuses.com/429
                title: Request Error
                status: 429
                detail: Rate limit exceeded. Try again later.
                instance: /api/courses
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            example:
              value:
                type: https://httpstatuses.com/404
                title: Request Error
                status: 404
                detail: Resource not found
                instance: /api/courses/unknown
    ValidationError:
      description: Validation error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            example:
              value:
                type: https://httpstatuses.com/400
                title: Request Error
                status: 400
                detail: name is required (string)
                instance: /api/courses
  schemas:
    Course:
      description: Course within an organization.
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        code:
          type: [string, 'null']
      required: [id, name]
    CourseListItem:
      description: Course item as returned in list endpoints (includes createdAt)
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        code:
          type: [string, 'null']
        createdAt: { type: string, format: date-time }
      required: [id, name, createdAt]
    CreateCourseInput:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
      required: [name]
    UpdateCourseInput:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
...
    PlanSessionOut:
      description: Planned study session returned by planner.
      type: object
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        topic:
          type: [string, 'null']
        description:
          type: [string, 'null']
        status: { type: string }
    PlanSessionsResponse:
      description: Planner result with plan id and generated sessions.
      type: object
      properties:
        planId: { type: string }
        sessions:
          type: array
          items: { $ref: '#/components/schemas/PlanSessionOut' }
      required: [planId, sessions]
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
paths:
...
  /api/courses:
    get:
      summary: List courses in caller's org
      tags: [Courses]
      security: [ { bearerAuth: [] } ]
      operationId: listCourses
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/CourseListItem' }
              examples:
                sample:
                  value:
                    items:
                      - { id: "c_123", name: "Intro to Databases", code: "DB101", createdAt: "2025-09-01T10:00:00.000Z" }
                      - { id: "c_124", name: "Algorithms", code: "CS201", createdAt: "2025-09-03T08:30:00.000Z" }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      summary: Create a course in caller's org (instructor/admin)
      tags: [Courses]
      security: [ { bearerAuth: [] } ]
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCourseInput' }
            examples:
              valid:
                value:
                  name: "Intro to Databases"
                  code: "DB101"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }
              examples:
                created:
                  value: { id: "c_123", name: "Intro to Databases", code: "DB101" }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { description: Forbidden }
        '409': { description: Conflict }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/courses/{courseId}:
    patch:
      summary: Update a course (instructor/admin)
      tags: [Courses]
      security: [ { bearerAuth: [] } ]
      operationId: updateCourse
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCourseInput' }
            examples:
              rename:
                value:
                  name: "Algorithms I"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }
              examples:
                updated:
                  value: { id: "c_124", name: "Algorithms I", code: "CS201" }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { description: Forbidden }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete a course (instructor/admin)
      tags: [Courses]
      security: [ { bearerAuth: [] } ]
      operationId: deleteCourse
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: No Content }
        '403': { description: Forbidden }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '409': { description: Conflict }
...
  /api/planner/plan:
    post:
      summary: Generate plan sessions from constraints
      tags: [Planner]
      security: [ { bearerAuth: [] } ]
      operationId: generatePlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fromDate: { type: string, format: date-time }
                toDate: { type: string, format: date-time }
                sessionMinutes: { type: integer }
                dailyCap: { type: integer }
                preferredStartHour: { type: integer, minimum: 0, maximum: 23 }
                preferredEndHour: { type: integer, minimum: 0, maximum: 23 }
                topics:
                  type: array
                  items: { type: string }
                description: { type: string }
              required: [fromDate, toDate, sessionMinutes, dailyCap]
            examples:
              basic:
                value:
                  fromDate: "2025-09-22T00:00:00.000Z"
                  toDate: "2025-09-28T23:59:59.000Z"
                  sessionMinutes: 60
                  dailyCap: 3
                  preferredStartHour: 9
                  preferredEndHour: 18
                  topics: ["Databases", "Algorithms"]
                  description: "Midterm prep"
      responses:
        '201':
          description: Created with sessions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanSessionsResponse' }
              examples:
                sample:
                  value:
                    planId: "p_001"
                    sessions:
                      - { start: "2025-09-22T09:00:00.000Z", end: "2025-09-22T10:00:00.000Z", topic: "Databases", description: "Normalization", status: "scheduled" }
                      - { start: "2025-09-22T11:00:00.000Z", end: "2025-09-22T12:00:00.000Z", topic: "Algorithms", description: null, status: "scheduled" }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { description: Conflict (overlaps) }
        '500': { description: Internal Server Error }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...